requirements.txt
torch>=2.0
torch-geometric>=2.4
stix2>=3.0
taxii2-client>=2.3
misp-stix>=2.4
sentence-transformers>=2.2
scikit-learn>=1.3
umap-learn>=0.5
matplotlib>=3.7
hdbscan>=0.8
pyyaml>=6.0
huggingface_hub>=0.20
